Move Credentials to own file
